fix: rename Conteact component directory to Contact
